+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-20T12:03:50.114Z",
+    "2026-02-20T13:15:46.741Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {