+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-20T12:03:50.114Z",
+    "2026-02-20T13:15:46.741Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {